Add SafeNetOnboardingRequiredAction class (v1.1)
This commit adds the SafeNetOnboardingRequiredAction class, which is responsible for handling a required action related to SafeNet MFA onboarding. The class implements the RequiredActionProvider and RequiredActionFactory interfaces.

The SafeNetOnboardingRequiredAction class evaluates triggers and checks if the user has already been onboarded. If not, it adds the required action ID for SafeNet MFA onboarding.

When the required action is challenged, the class creates a response for updating the password.

In the process action method, the class performs the necessary steps for updating the user's password. It verifies the password and updates it, logs out other user sessions, sets the password set flag, and removes the default "Update Password" action if it exists. Finally, it prompts the user to re-authenticate.

This commit ensures that the SafeNet MFA onboarding action replaces the default "Update Password" action on the first login attempt, that is while the user does not have the 'password-set' flag attribute, providing a more specialized functionality for password management in relation to third-party identity providers.

Version bumped to 1.1 to indicate the addition of the SafeNet MFA onboarding functionality.